requirements.txt
flask==2.3.3
pinecone==7.3.0
requests>=2.32.3
python-docx==0.8.11
PyPDF2==3.0.1
python-dotenv==1.0.0
huggingface_hub==0.16.4
gunicorn==21.2.0
werkzeug>=2.3.0
blinker>=1.6.0
click>=8.0.0
colorama>=0.4.6
itsdangerous>=2.1.0
jinja2>=3.1.0
markupsafe>=2.1.0